+++ GlobalUsings.cs
@@ -10,6 +10,7 @@ global using System.Diagnostics;
 global using System.Diagnostics.CodeAnalysis;
 global using System.Globalization;
 global using System.IO.Compression;
+global using System.IO.MemoryMappedFiles;
 global using System.Net;
 global using System.Net.Http.Headers;
 global using System.Net.NetworkInformation;
