+++ src/Sentry/Internal/ILSpy/SingleFileBundle.cs
@@ -1,14 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
-using System;
-using System.Collections.Immutable;
-using System.IO;
-using System.IO.MemoryMappedFiles;
-using System.Runtime.CompilerServices;
-using System.Text;
-using System.Xml.Linq;
 
-namespace Sentry.Internal.ILSply;
+namespace Sentry.Internal.ILSpy;
 
 #if NETCOREAPP3_0_OR_GREATER && PLATFORM_NEUTRAL
